web_ui/requirements.txt
Flask==2.2.3
Flask-WTF==1.1.1
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-RESTful==0.3.10
Flask-Cors==3.0.10
gunicorn==20.1.0
requests==2.28.2
python-dotenv==1.0.0
PyJWT==2.6.0
bcrypt==4.0.1
Werkzeug==2.2.3
plotly==5.14.1
pandas==2.0.0
numpy==1.24.3
